src/requirements.txt
streamlit>=1.32.0
azure-identity>=1.15.0
azure-ai-projects==1.0.0b5
openai>=1.12.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
msal>=1.24.0
requests>=2.31.0
